/ppbox/branches/1.2.0/src/ppbox/avformat/mp4 修正 mp4 格式 emailto: box-dev
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@8028 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ mkv/ebml/EBML_Archive.h
@@ -226,7 +226,7 @@ namespace ppbox
                         Element & elem2 = elem.items[i];
                         ar_.seekg(elem2.offset + elem2.head_size(), std::ios::beg);
                         (EBML_ElementHeader &)v = elem2;
-                        _Traits::load(ar_, v.value(), elem2.Size);
+                        _Traits::load(ar_, v.value(), (size_t)elem2.Size);
                     }
                 }
             }
